Move template update coordinator to its own file
